Add promote_secondaries keyword for vrrp_instance block
If two IPv4 VIP addresses are in the same CIDR, and the primary
address is removed, then by default any other address in the same CIDR is
also removed. To stop this happening, the promote_secondaries flag
needs to be set on the interface.

Commit e5526cf added setting the promote_secondaries option on
VMAC interfaces, and stated that adding the option for non-VMAC
interfaces would be added later. This commit now adds a
promote_secondaries configuration option in order to set the flag
on the interface.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>