Add conditional configuration feature
It is usually the case that the configurations for keepalived for
systems operating together are virtually identical, and only differ
in vrrp instance priorities, router id, and unicast addresses if
those are being used.

It is a nuisance to have to edit one file for each server to make
identical changes, so this commit adds the facility for conditional
configuration entries.

Any line starting with the '@' character is a conditional line.
Immediately following the '@' character is a config id. The line is
only included in the configuration if the config id matches the
argument passed to keepalived with the -i option on the command line.

For example, consider the following configuration snippet:

global_defs
{
@main	router_id main_router
@backup router_id backup_router
}

If keepalived is started with -i main, then the router id will be
main_router, if started with -i backup, then backup_router. If
keepalived is started without the -i option, or -i anything else,
then the above snippet will not configure any router id.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>