Stop segfaulting when configuration keyword is missing its parameter
There are many places where during configuration parsing the code
assumes that if a keyword is specified that requires a parameter, then
the parameter exists. If the parameter doesn't exist, then the code
indexes past the end of the vector, and at best segfaults, and at worst
may carry on, parsing random data.

This commit adds strvec_slot() which checks for the presence of the
parameter, and if configured will call a function that can handle the
error. Currently this logs that the parameter is missing, with as much
helpful information as it can provide, and then terminates.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>