Correctly handle existing VMACs on reload
Anthony Dempsey <adempsey@brocade.com> in issue #449 identified that
keepalived attempts to recreate existing VMAC interfaces on a reload,
and that the subsequent failure causes keepalived not to use the VMAC.

This then identified further issues such as the check for an existing
VMAC in netlink_link_add_vmac() didn't also check the interface a
VMAC was on, and that the checks for conflicts of VMAC interface names
with existing interfaces weren't sufficient.

This patch builds on the patch provided by Anthony Dempsey to also
resolve the additional issues identified.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>