Fix check of matching VRRP instances on reload
On a reload, clear_diff_vrrp() removes vrrp instances that are no
longer in the configuration. The check, however, was based on vrrp
instance name, which might have changed. The check is now based on
VRID, address family and interface, since it is this triplet that
uniquely defines a vrrp instance.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>