Ensure GARPs/GNAs are sent after reload if VIP/eVIP addresses added
Although there have been versions of keepalived when GARPs/GNAs were
sent after a reload, this was due to a bug in determining if the VRRP
instance had existed before. Resolving that bug (commit aaea8a4),
caused keepalived to stop sending GARPs after a reload. This commit
now specifically adds code to send GARPs on a VRRP instance for all
addresses on that instance. It would be better if GARPs were sent only
for the added addresses, and that may be resolved in a future commit.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>