Resolve "Netlink: error: message truncated" messages
On systems with a page size larger than 4096 keepalived may report:

"Netlink: error: message truncated" messages

This error was reported on a ppc64le in an OpenStack/Nutron environment.
Ppc64le is using a 64k pages size. I found that keepalived's netlink recvmsg
buffer was too small causing messages to be truncated. The size of the read
buffer for the netlink socket should be based on page size however, it should
not exceed 8192. See the comment in the patch.

I tested the fix by creating 100 veth interfaces and verifying the errors
did not return.

Signed-off-by: David Wilder <dwilder@us.ibm.com>
Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>