Use ipsets with namespaces on Linux < 3.13 if ipset names configured
The problem with using ipsets with namespaces on Linux < 3.13 is that
ipsets were not namespace aware, and so all ipset entries added are
global to the system, including all network namespaces. This causes
problems if the default ipset names are used, but if set names have
been specified, it is reasonable that they have been set to be
different for each namespace, and hence there will be no clashes.

The documentation is also updated for vrrp_ipsets keyword.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>