Don't segfault if modules ip_tables or ip6_tables not loaded
If either of the modules is not loaded, then don't use ip(6)tables for
that address family. We could load the module, but there would be no
entries pointing to the chains that we use, and so there is no point
adding entries to chains that won't be traversed.

This resolves issue #457

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>