Don't chdir("/") if not forking
In keepalived_main() there is a comment that the working directory
is / unless keepalived is run in non-forked mode, in which case it
remains the current working directory when keepalived was run.

Unfortunately start_vrrp_child() and start_check_child() were
executing chdir("/") regardless of whether they had been forked or not.

Since the parent process does chdir("/") if it is appropriate, the
children will inherit that, so they don't need to chdir() at all.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>