Use nodename as default id for conditional configuration
If the node name returned by uname() is host123.abc.de, then
lines in the configuration file matching @host123 will match the
conditional configuration test. This means that it is no longer
necessary to specify the -i command line option if the conditional
configuration string used in the configuration is the node name.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>