Use execve() rather than system() for executing scripts
execve() saves a fork and an extra process running while the script
is executing. It also allows the parameters to be pre-parsed into
an array, and ensures that there is no path search when the script
is executed. It will also allow PR_SET_PDEATHSIG to be set for the
script process.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>