Add support for using hot-swappable NICs
All configuration data now uses an interface_t * for any reference to
an interface. When an interface is deleted, the interface_t structure
is updated to set ifindex = 0. When an interface is created, the
ifindex in the interface_t structure is updated. It is therefore
important that there is no place other than in the interface_t structures
where the ifindex is held.

When the configuration is read, if an interface is specified that doesn't
already exist, a new interface_t structure is added to the interface list,
with only the interface name in the structure. This allows configuration to
point to the relevant interface even though it may not exist. If a
non-existent interface is specified, a warning is logged, since it may be
a configuration error, or it may be that the interface doesn't exist at the
time keepalived starts.

To be implemented:

When an interface is deleted, any route or ip address that has any
configuration related to that interface is also deleted. This applies to
static addresses, static routes, virtual routes and vips/evips that
are configured to interfaces other than the one the vrrp instance
is running on.
NOTE: static and virtual rules are not deleted when a relevant interface
is deleted.

This actually isn't a problem relating to interfaces being deleted, but
the address/route deletion occurs when an interface goes down, either
because it is configured down, or for example if a network cable is
disconnected, so this is a problem that already exists in keepalived.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>