ipwrapper.c: make functions void if return value not used or constant
Several functions in check/ipwrapper.c were always returnung the same
value, and the code calling the functions then checked and returned an
error if the return value was not the value always returned. Also, for
some functions returning a value the return value was never checked in
the calling function.

Making the functions void, and removing the if (...) makes the code easier
to read, and potentially slightly more efficient.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>