Make builds reproducable, and copyright date reflect latest commit
Pull request #503 provided an update to facilitate reproducable builds,
and also ensure that the copyright date doesn't postdate the last source
code modification. Unfortunately the commits required manual updates to
change the copyright year, thereby creating maintainability issues. The
commit also allowed fake build dates to be specified.

This commit takes an alternative approach, and takes the dates used for
the copyright message and the version date from the date of the last git
commit.

If the code is build from within a git tree, this is straightforward. On
the other hand, the code may be build from  tarball, so we ensure that
the lib/git-commit.h file is updated when the tarball is built, and
included in the tarball.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>