Log errors if configure IPVS with IPv6 if not using libnl
The socket interface for configurating IPVS does not support IPv6,
so rather than leaving the user with the error message
"Operation not supported with specified address family"
give a meaningful message in the log.

At configure time, a warning will also be generated stating that
IPVS will not support IPv6.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>