Log if virtual_server_group doesn't exists, or address family mismatch
If a virtual server is configured to use a virtual server group but that
virtual server group doesn't exist, then log an error.

Also, if the address family of the virtual server group and virtual server
don't match, log an error as well.

We really ought to be removing the virtual server from the configuration,
but I haven't worked out how to do that yet.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>