Don't flag changes to automake/conf generated files as source changes
The output of keepalived -v adds a '+' if there are uncommitted changes to
the source code. However, we aren't interested in changes to the autoconf
or automake generated files, since these aren't really "source" files, and
are only included in the git repo to allow building on systems without
autoconf/automake. Further, the differences may simply be due to different
versions of autotools being used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>