Force recreating automake/autoconf files when building with rpm
If an autoconf/automake source file is patched as part of the rpmbuild
process, then some of the autoconf/automake generated files may be
regenerated, and this can cause a mismatch if the versions of autoconf/
automake on the system building the rpm don't match the versions
that were used to generate the files that have been committed to git.

This patch changes the keepalived.spec file to always run autoreconf -f -i
to ensure the generated autoconf/automake files are aligned to the right
version.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>