Revert and fix of last bad commit.
This reverts commit 73929816d79c993f9a36edf2575d02d19958e324.

priority should be given to priority owner case not to the pre-emption
mode(As per the RFC).

Check for preemption only if its not the case of priority owner.

In case one router having preemption enabled and it is acting as backup, but if we
configure the same router as priority owner router in that case this should become master irrespective of the preempt state.