Ensure nopreempt is not set if address owner
Pull request #523 identified that preemption should always apply
if a vrrp instance is the address owner. This commit takes the idea
of the patch in that pull request, but applies it at config time,
logs a warning if nopreempt is requested for an address owner, and
clears the nopreempt flag.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>