Don't regenerate VRRP packets every time they are sent
Each vrrp instance has it's own send buffer, so let's make use of
that and not re-write the entire packet each time, but just update
the fields that need changing, which of course includes the checksums.

In the course of developing this update it was discovered that there
is no need to generate the IP header checksum at all, since the kernel
will always generate it.

When using unicast peers, the ip header id field, and the ipsecah
sequence number don't need to be updated for each packet sent, but only
once for each time we send a block of updates to all the peers.

RFC1624 documents how to update the internet checksum as values are
changed, which is significantly quicker than regenerating the checksum
over all the data.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>