Make ha_suspend work when building without vrrp
Support of ha_suspend was only enabled when keepalived was built
with vrrp support. There may be other processes that are adding and
deleting ip addresses, so support of ha_suspend should be enabled when
building without vrrp support.

Also, the vrrp process doesn't need to call the update_checker_activity()
function when addresses are added or deleted.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>