check/libipvs: add missing include
u_int16_t et al. are from stdint.h, so it needs to be included.

It happens to work on glibc and uClibc, becasue those C libraries are
very much lax on what headers they internally include, so stdint.h is
pulled in automagically.

However, musl is very conservative and standards-compliant, so its
headers do not internally include much.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>