Dynamic/static linking options of libnl/libnl-3, libip[46]tc and libipset
libnl/libnl-3, libip[46]tc and libipset can all be dynamically linked at
run-time, and if they are not available, keepalvied will use the alternative
code which is used when the libraries cannot be linked a build time.

This means that a single executable keepalived can be created that will use
the libraries if they are installed on the target system, but will fall back
to the alternatives if the libraries are not available.

This is useful for build environments such as Buildroot which will not force
optional dependencies (see pull request #540), since now keepalived can be
built so as not to force the optional dependencies, but to make use of them
if they are installed.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>