Fix creation of iptables entries on more recent kernels
On a 4.9.13 kernel iptables entries were being created with
return-nomatch ! update-counters ! update-subcounters, as shown by
the iptables command.

Although it is not understood why these options are being added, it
transpires that the problem occurs when using version one of the
xt_info_set_match, but doesn't occur when using version 4 of the
structure.

This patch ensures that the latest version of the structure that is
supported by the kernel is always used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>