Ensure locations of pid files is consistent
Issue #563 identified that the generated keepalived.service has the
wrong location for the pid file. On investigating this it was discovered
that keepalived isn't following the GNU coding standards for location of
pid files; however, we can't now move the default location of pid files.

This commit ensures that the keepalived.service file's location for pid
files is consistent with where keepalived is placing them, but also adds
a configure option --enable-gnu-std-paths, which means that keepalived
will use ${localstatedir} for the location of pid files, while the default
remains /var/run

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>