Don't segfault if unable to load ip_vs module
In a docker container it isn't possible to load a kernel module. The
check code was detecting that it couldn't load the module, but the
checker process, when cleaning up prior to exiting, was assuming that
certain pointers had been initialised which hadn't been when an error
was detected so early in the initialisation.

This commit adds testing for uninitialised pointers during the exit
sequence.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>