Handle missing virtual server configuration
keepalived was segfaulting if a virtual server had no real servers
configured. There were also issues of checkers running even if there
was missing essential configuration from a virtual server which meant
it could be set up. The problems were a virtual server group specified
but it didn't exist, a virtual server group with no configuration, and
a virtual server address family not match the address family of a virtual
server group.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>