Add some additional docker support files and add make target docker
The configuration file installed by make install isn't ideal to run
keepalived with, so add a simple keepalived.conf that will be
installed into the container.

Add make target docker, to build the docker image.

Add docker/README to give some information about building and using
containers (this is mainly so that I don't forget how to the details).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>