Fix startup with tracking scripts
We need to know how many tracking scripts and for sync group
how many of their instances are in init state, pending the first
completion of scripts.

The effect of this change is that vrrp instances won't come up and
then promptly go down again after the first run of the script has
produced a fault.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>