Only delete virtual server once if using a virtual server group
If multiple virtual servers are using a virtual server group, the
virtual servers are defined by the virtual server group, and so they
should only be deleted for the first virtual server using the group.

There is still an issue that the configuration of all virtual servers
using the virtual server group needs to be consistent.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>