Add lvs_method per real server
The lvs_method should be settable for each real server within a
virtual server. This commit maintains existing default behaviour
by using the lvs_method set against the virtual server as the default
for the real servers, but adds the option to configure the lvs_method
individually for each real server.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>