Add vrrp_track_file option to read priority change from a file
Add option to monitor files which contain a numeric value which
is added to effective_priority, or puts a vrrp instance into failed state.
This is an alternative to track scripts and is more efficient, more
flexible and more responsive:

i)   Nothing has to be done on a periodic basis; the monitored file is only
     read when it is changed
ii)  The change in priority is based on the value in the file, rather than
     preconfigured for track scripts
iii) The change in priority is processed as soon as the file is changed,
     rather than waiting for the next interval to expire to run the script.

This coupled with notify FIFOs significantly reduces the need to run scripts,
which can have a large overhead in processing time.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>