Always make a vmac interface point to it's base interface
This commit changes the internal behaviour of vmacs. Previously,
if the underlying interface of the vmac was deleted, the vmac
interface structure was updated to make base_ifp point back to
the vmac interface structure. Now, the vmac interface structure
continues pointing at the underlying interface, but the ifindex
is set to 0 if the interface doesn't exist.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>