Fix checksum calculation for VRRPv3 IPv4 unicast
Issue 642 identified that Wireshark was reporting incorrect checksums
when using unicast peers with VRRPv3 and IPv4.

Commit 67275d2 resolved an issue of inconsistent calculation of
checksums (checksums in alternate packets were 0).

Wireshark was reporting incorrect checksums due to the IP address
being put in the pseudo IPv4 header for the calculation of the VRRP
checksum being 224.0.0.18. For unicast peers the address should be
the actual destination IP address. Since keepalived was doing the
same thing when a packet was received, it did not detect an incorrect
checksum.

There is now an issue of backward compatibility of keepalived prior to
this commit with keepalived from this commit onwards when using VRRPv3
with unicasting. Commit 67275d2 enabled VRRPv3 with more than one
unicast peer to work for the first time since VRRPv3 was introduced by
commit 02e1959. We can conclude therefore that no-one has been using
VRRPv3 with more than one unicast peer.

The next commit will introduce a mechanism to allow backward compatibility.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>