Make checksum change backwards compatible
This commits adds the ability of keepalived to revert to using the old
checksum calculation if it sees an advert that has an old checksum. This
means that if an old and a new version of keepalived are working
together, once the new version has seen an advert from an old version,
it will drop back to using the old style checksum, and so the two
keepalived instances will work together.

There is a slight problem with this, though. If the old version starts
when a new version is master, if will report bad checksums. This should
be fine, since keepalived should discard the adverts, time out and send
an advert which would make the new keepalived revert to old checksums.
Unfortunately, keepalived does not completely ignore bad adverts, since
it resets its master down timer, even for bad adverts, and so it never
times out. However, in this scenario, there will still remain one master
and all the other keepalived instances will be in backup state, and so
VRRP functionality is preserved.

As identified in commit bcf2936 until commit 67275d2 keepalived did not
work with VRRPv3 and more than one other unicast peer, so for migration
we only need to consider two unicast peers. To upgrade, first upgrade
the keepalived instance that is in backup mode. This will see old
checksums when it starts, and so start using old checksums. The other
keepalived instance can then be upgraded and it will also see old
checksums when it starts up. In order to switch to using new checksums,
temporarily add the following line in the configuration of each vrrp
instance that is in master state:

old_unicast_checksum never

and then reload the keepalived instance by sending it SIGHUP. After the
master has restarted, restart the backup with a SIGHUP, and they will
then be using new checksums. The temporary old_unicast_checksum never
lines can now be removed from the configuration.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>