Improve handling of virtual server groups
If multiple virtual servers use the same virtual server group, and
the virtual servers have different protocols, or the virtual server
groups are defined using only fwmarks and the virtual servers have
different address families, then multiple versions of the entries
in the virtual server groups will need to be created as IPVS
virtual servers.

This patch handles the creation and removal of the necessary different
virtual servers for the virtual server groups.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>