parser: do not exit when glob() doesn't match any files
Commit 3455c4a9 changed read_conf_file()'s behaviour to exit when glob()
returned non-success. Calling exit(3) from within a utility function
like read_conf_file causes a number of issues:

 - Keepalived will refuse to start when used together with `include` to
   modularize configuration (e.g. include
   /etc/keepalived/puppet.d/*.conf which may be empty).
 - It also causes the main process to exit uncleanly during reloads,
   leaving orphan child processes around. This issue is serious, as it
   leaves keepalived in a crippled state; subsequent attempts to restart
   the keepalived service without manually killing the VRRP/healthcheck
   children first will fail.

The original commit message makes it clear that this is about the main
configuration file, which is essential indeed. However, the presence of
the main config file is also checked and handled by check_conf_file() on
startup.

Revert to the previous behavior by having read_conf_file() log glob()
failures and skip using the glob buffer. For more fine-grained control
of file read failures, the signature of read_conf_file() should be
changed to return a value that can be used to check for and handle
errors.