Don't increment local ipsecah_counter for every received pkt
If an ipsecah advert is received, we don't want to increment our
copy of the counter until the packet is fully validated. In particular,
if the counter is one too low in the received packet, if we increment
the counter every time a packet is received, we will never re-sync
with the new master. Also, if there has been a split-brain situation
causing there to be two masters, when we start seeing both, and until
one of them drops back to backup, we will keep incrementing our counter
for each advert, the counter will get ahead, and again we won't be able
to recover.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>