Den't decrement ipsecah_counter when transition to backup
Since commit 4a7da48 stopped incrementing the counter for every
advert received, we no longer need to decrement the counter on
transition to backup. The need previously to decrement was due to
a small timing window which could have caused a duplicate sequence
number to be received, and since we incremented our counter every
time a packet was received, we could never resync. Now, we will
just discard a packet with a duplicate sequence number, and the
next packet will have an acceptable sequence number.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>