Don't send pri 0 adverts when syncing members of sync group to backup
If a system is transitioning a member of a sync group to master, it
has determined that all members of the sync group will become master.
Therefore, when the previous master receives the first higher priority
advert for the first vrrp instance, it will have to sync the other
members of the sync group to backup, but since the new master has
already decided it will become master for all vrrp instances in the
group, there is no benefit in the old master sending prio 0 adverts
when the other members of the sync group transition to backup, since
a new master is already known for those instances.

This also stops a problem with AH authentication when this happens.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>