Add weights to track_files & track_script/files/iface to sync groups
Allow specifying a weight for a track file. This multiplies the value
read from the tracked file to give the weight change.

Add track_script, track_interface and track_file to vrrp sync groups.
The reason for this is that it allows weights to be specified against
the track items and guarantees that the same weights are configured
for all vrrp instances in the sync group. This obviates the need for
global_tracking, which not only means that the tracked items don't need
to be configured against each vrrp instance in the sync group, but also
ensures that the weights applied are all the same.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>