Deprecate global_tracking and add sync_group_tracking_weight
Now that sync groups support track_script, track_interface and
track_file, there is no need for global_tracking. However,
with global_tracking it was possible to configure different
tracking weights for the same tracking object on difference
vrrp instances within a sync group, so we add
sync_group_tracking_weight to allow that to continue.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>