Fix not allowing tracking weights for sync groups and address owner
Tracking configured on a vrrp instance which is a member of a sync
group must not be weighted unless sync_group_track_weighted is set.
Likewise, an address owner must not have weighted tracking.

Previously, after a tracking object updated its status, a check was
made if the vrrp instance was the address owner. This is now changed
so that at startup, an address owner is set not to have any tracking
weights, and if appropriate the vrrp instances stops tracking that
object.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>