Add high performace child finder code
The code to find the relevant thread to execute afer a child process
(either a vrrp track script or a misc_check healthchecker) was doing
a linear search for the matching pid, which if there are a large number
of child processes running could become time consuming.

The code now will enable high performance child finding, based on using
mlists hashed by the pid, if there are 32 or more vrrp track scripts or
misc check healthcheckers. The size of the mlist is based on the number
of scripts, with a limit of 256.

This is from code written by Damien Clabaut <damien.clabaut@corp.ovh.com>

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>