Allow tunnelled rs address family not to match vs family
The address family of a tunnelled real server does not have to
match the address family of its virtual server, so we need to
delay any setting of the vs address family from an rs address
until the end of the real_server block, so that we know whether
the forwarding method is tunnelling or not. Likewise the check
of the sorry server has to be delayed until the end of the
virtual server configuration (the tunnelling method may be
specified after the address of the real/sorry server).

The address family of a virtual server is only not determined
by the virtual server configuration itself if the virtual server
is defined by a fwmark and all of the real/sorry servers are
tunnelled. In this case the address family cannot properly be
determined from the address family of any tunnelled real servers.
However, to maintain backward compatibility with configurations
used prior to this commit, the address family of the virtual
server will be taken from the address family of the (tunnelled)
real/sorry servers if they are all the same; if they are not all
the same it will default to IPv4 (this is not incompatible since
previously mixed IPv4 and IPv6 real/sorry servers were not allowed,
even if tunnelled).

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>