Fix removing left-over addresses if keepalived aborts
Issue #718 reported that if keepalived terminates abnormally when
it has vrrp instances in master state, it doesn't remove the
left-over VIPs and eVIPs when it restarts. This is despite
commit f4c10426c saying that it resolved this problem.

It turns out that commit f4c10426c did resolve the problem for VIPs
or eVIPs, although it did resolve the issue for iptables and ipset
configuration.

This commit now really resolves the problem, and residual VIPs and
eVIPs are removed at startup.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>