Fix use of init_state after a reload
Issue #627 identified that vrrp->init_state was being incorrectly
used in vrrp_fault(), since it is modified at a reload.

Instead of using init_state, we now use the configured priority
of the vrrp instance, so if the vrrp instance is the address owner
(priority 255) it will transition to master after leaving to fault
state, otherwise it transitions to backup.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>