Fix building with LibreSSL version of OpenSSL
Unfortunately LibreSSL updates OPENSSL_VERSION_NUMBER, and its value
is higher that OpenSSL's latest version. When checking the version
number we need to check that we are not using LibreSSL (by checking
whether LIBRESSL_VERSION_NUMBER is defined).

LibreSSL also hasn't implemented the new functions that OpenSSL has
provided to replace functions that are deprecated or it is recommended
should not be used, and so if using LibreSSL the old functions need
to be used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>